requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.8.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
